Add MoneyInputLayout, migrate send and receive
Consolidate useMoneyInputField, MoneyInputLayout, and the display
components into a single money-input-layout module. The hook and layout
are always used together so they live in one file. Display components
are now internal — consumers compose via the layout's slots (children,
actions, belowDisplay).

Send uses belowDisplay for its destination row between the amount
display and account selector.
